server/requirements.txt
flask
flask-cors
numpy
pandas
tensorflow-cpu
scikit-learn
joblib
requests
openpyxl
gunicorn